requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Celery and Redis
celery==5.3.4
redis==5.0.1

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# AI/ML Models
torch==2.1.1
transformers==4.36.0
sentencepiece==0.1.99
protobuf==4.25.1

# Speech Processing (✅ compatible with Python 3.10)
openai-whisper==20231117
TTS==0.17.5
soundfile==0.12.1
librosa==0.8.1
numpy==1.22.0

# Language Detection
langdetect==1.0.9

# Evaluation Metrics
sacrebleu==2.3.1

# Monitoring and Logging
prometheus-client==0.19.0
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
